decouple auth with useAuthService Hook
